python/requirements.txt
numpy~=1.0
scipy
cupy-cuda12x
tqdm
